pythonProject10/employee/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-extensions==3.1.5
django-utils==0.0.2
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.6
utils==1.0.1
